Update app structure and dependencies; move AboutUs component to tabs
