+++ package.json
@@ -18,10 +18,10 @@
     "@expo/vector-icons": "^14.0.2",
     "@react-navigation/bottom-tabs": "^7.2.0",
     "@react-navigation/native": "^7.0.14",
-    "expo": "~52.0.35",
+    "expo": "52.0.36",
     "expo-blur": "~14.0.3",
-    "expo-constants": "~17.0.6",
-    "expo-font": "~13.0.3",
+    "expo-constants": "17.0.7",
+    "expo-font": "13.0.4",
     "expo-haptics": "~14.0.1",
     "expo-linking": "~7.0.5",
     "expo-router": "~4.0.17",
@@ -38,7 +38,8 @@
     "react-native-safe-area-context": "4.12.0",
     "react-native-screens": "~4.4.0",
     "react-native-web": "~0.19.13",
-    "react-native-webview": "13.12.5"
+    "react-native-webview": "13.12.5",
+    "jest-expo": "52.0.5"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
